Split out pycore_crossinterp_data_registry.h.
